Multiple databases
Databases are opened when the user uses the ":db" or ":open" commands, and are closed when the user either uses the ":close" or the ":quit" command.
Errors are displayed if the user tries to run a Daison statement without first opening a database, or if they try to close a database which has not yet been opened.